move tracking function
