+++ src/evaluator.ts
@@ -497,53 +497,6 @@ export function generateVarCombinations(
   return combinations;
 }
 
-// Add this function to display the token usage summary
-function displayTokenUsageSummary() {
-  const tracker = TokenUsageTracker.getInstance();
-  const providers = tracker.getProviderIds();
-
-  if (providers.length === 0) {
-    return;
-  }
-
-  logger.info('\n==== Token Usage Summary ====');
-
-  // Display per-provider usage
-  providers.forEach((providerId) => {
-    const usage = tracker.getProviderUsage(providerId);
-    if (!usage) {
-      return;
-    }
-
-    const promptTokens = usage.prompt || 0;
-    const completionTokens = usage.completion || 0;
-    const cachedTokens = usage.cached || 0;
-    const totalTokens = usage.total || promptTokens + completionTokens;
-    const numRequests = usage.numRequests || 0;
-
-    logger.info(
-      `${providerId}: ${totalTokens.toLocaleString()} tokens total ` +
-        `(${promptTokens.toLocaleString()} prompt, ${completionTokens.toLocaleString()} completion, ` +
-        `${cachedTokens.toLocaleString()} cached) across ${numRequests} requests`,
-    );
-  });
-
-  // Display total usage across all providers
-  const totalUsage = tracker.getTotalUsage();
-  const totalPrompt = totalUsage.prompt || 0;
-  const totalCompletion = totalUsage.completion || 0;
-  const totalCached = totalUsage.cached || 0;
-  const grandTotal = totalUsage.total || totalPrompt + totalCompletion;
-  const totalRequests = totalUsage.numRequests || 0;
-
-  logger.info(
-    `\nTotal: ${grandTotal.toLocaleString()} tokens ` +
-      `(${totalPrompt.toLocaleString()} prompt, ${totalCompletion.toLocaleString()} completion, ` +
-      `${totalCached.toLocaleString()} cached) across ${totalRequests} requests`,
-  );
-  logger.info('==============================\n');
-}
-
 class Evaluator {
   evalRecord: Eval;
   testSuite: TestSuite;
@@ -1428,9 +1381,6 @@ class Evaluator {
       suite: testSuite,
     });
 
-    // Display token usage summary
-    displayTokenUsageSummary();
-
     telemetry.record('eval_ran', {
       numPrompts: prompts.length,
       numTests: prompts.reduce(
@@ -1458,6 +1408,8 @@ class Evaluator {
       hasAnyPass: this.evalRecord.prompts.some(
         (p) => p.metrics?.testPassCount && p.metrics.testPassCount > 0,
       ),
+      // FIXME(ian): Does this work?  I think redteam is only on the config, not testSuite.
+      // isRedteam: Boolean(testSuite.redteam),
     });
 
     // Update database signal file after all results are written
